+++ docs/CHANGELOG.md
@@ -10,6 +10,14 @@ Known [issues](https://github.com/mgerhardy/vengi/issues?q=is%3Aissue+is%3Aopen+
 
 ## 0.0.27 (2023-XX-XX)
 
+Breaking changes:
+
+   - Renamed globals for lua scripts. Added a `g_` prefix to them. You have to adopt your scripts
+     to work the latest version if you use them
+     * `scenegraph` is now `g_scenegraph`
+     * `palettemgr` is now `g_palette`
+     * `noise` is now `g_noise`
+
 General:
 
    - Added cvar `voxformat_pointcloudsize` for point cloud formats